requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argcomplete==3.5.3
asgiref==3.8.1
attr==0.3.1
attrs==25.1.0
audioread==3.0.1
azure-core==1.32.0
azure-storage-blob==12.24.1
basicsr==1.4.2
black==25.1.0
bleach==5.0.1
blinker==1.9.0
boto==2.49.0
boto3==1.36.26
botocore==1.36.26
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==44.0.1
cycler==0.12.1
datamodel-code-generator==0.26.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
Django==4.2.19
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==2.4.0
django-migration-linter==5.1.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dotenv==0.9.9
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
exceptiongroup==1.2.2
expiringdict==1.2.2
Faker==36.1.1
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
fsspec==2025.2.0
future==1.0.0
genson==1.3.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-appengine-logging==1.6.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.2
google-cloud-logging==3.11.4
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
gradio==4.44.1
gradio_client==1.3.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.3.0
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflect==5.6.2
inflection==0.5.1
inplace-abn @ file:///D:/ml/tryon/inplace_abn
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
keras==3.8.0
kiwisolver==1.4.7
label-studio==1.15.0
label-studio-sdk==1.0.8
launchdarkly-server-sdk==8.2.1
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lmdb==1.6.2
lockfile==0.12.2
lxml==5.3.1
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
modules==1.0.0
mpmath==1.3.0
msgpack==1.1.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
ninja==1.11.1.3
nltk==3.9.1
noisereduce==3.0.3
numba==0.60.0
numpy==1.26.4
openai==1.64.0
opencv-python==4.11.0.86
opentelemetry-api==1.30.0
optree==0.14.1
ordered-set==4.0.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
pipenv==2024.4.1
platformdirs==4.3.6
pooch==1.8.2
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyboxen==1.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pyRFC3339==2.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.4
python-multipart==0.0.20
pytz==2022.7.1
PyYAML==6.0.2
redis==3.5.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
rich==13.9.4
rpds-py==0.23.1
rq==1.10.1
rsa==4.9
rstr==3.2.2
ruff==0.9.7
rules==3.4
s3transfer==0.11.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
semver==3.0.4
sentry-sdk==2.22.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
starlette==0.45.3
sympy==1.13.1
tb-nightly==2.20.0a20250224
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tifffile==2024.8.30
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchgeometry==0.1.2
torchvision==0.20.1+cu121
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
ultralytics==8.3.78
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
uvicorn==0.34.0
virtualenv==20.29.2
webencodings==0.5.1
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
xmljson==0.2.1
yapf==0.43.0
zipp==3.21.0